Documentación del proyecto

+++ database/migrations/2023_03_15_014204_create_accounts_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -13,7 +12,11 @@ return new class extends Migration
     {
         Schema::create('accounts', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
+            // Cuando se define un campo que va a hacer una llave foránea, debe tener el mismo tipo
+            // de dato y la misma longitud del campo a referencia
+            $table->unsignedBigInteger('user_id');
+            // A través del método ->foreign, le paso como parámetro, el campo que quiero agregarle la foránea
+            // y señalo a qué campo y de qué tabla pertenece dicha relación
             $table->foreign('user_id')->references('id')->on('users');
             $table->timestamps();
         });
